Simplify homepage and update README
