+++ README.md
@@ -2,16 +2,10 @@ This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next
 
 ## Getting Started
 
-First, run the development server:
+To run the development server locally, use the following command:
 
 ```bash
 npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
 ```
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
@@ -19,6 +13,27 @@ Open [http://localhost:3000](http://localhost:3000) with your browser to see the
 You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
 This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
+
+## Building and Serving Locally
+
+To build the project for production, run:
+
+```bash
+npm run build
+```
+
+Then, to serve the production build locally, run:
+
+```bash
+npm start
+```
+
+Additionally, to serve the production version locally using a static server, you can install `serve` (if you haven't already) and use it to serve your application:
+
+```bash
+npm install -g serve
+serve -s build
+```
 
 ## Learn More
